[Brownfield] Package rename (#1202)
Removed old react package and changed reactv2 to react.
