+++ android/app/src/main/java/com/microsoft/codepush/react/CodePushDialog.java
@@ -9,29 +9,52 @@ import com.facebook.react.bridge.LifecycleEventListener;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
-import com.microsoft.codepush.react.exceptions.CodePushUnknownException;
-import com.microsoft.codepush.react.utils.CodePushRNUtils;
+import com.microsoft.codepush.common.exceptions.CodePushGeneralException;
+import com.microsoft.codepush.common.interfaces.CodePushConfirmationCallback;
+import com.microsoft.codepush.common.interfaces.CodePushConfirmationDialog;
+import com.microsoft.codepush.common.utils.CodePushLogUtils;
 
-public class CodePushDialog extends ReactContextBaseJavaModule{
+import java.util.Arrays;
 
+/**
+ * Represents a react native dialog.
+ */
+public class CodePushDialog extends ReactContextBaseJavaModule implements CodePushConfirmationDialog {
+
+    /**
+     * Creates an instance of the {@link CodePushDialog}.
+     *
+     * @param reactContext instance of the {@link ReactApplicationContext}.
+     */
     public CodePushDialog(ReactApplicationContext reactContext) {
         super(reactContext);
     }
 
+    /**
+     * Displays dialog with the provided options.
+     *
+     * @param title              title of the dialog.
+     * @param message            message to be displayed.
+     * @param positiveButtonText text on the positive action button.
+     * @param negativeButtonText test on the negative action button.
+     * @param successCallback    callback to handle "OK" events.
+     * @param errorCallback      callback to handle "Discard" events.
+     */
     @ReactMethod
-    public void showDialog(final String title, final String message, final String button1Text,
-                           final String button2Text, final Callback successCallback, Callback errorCallback) {
+    public void showDialog(final String title, final String message, final String positiveButtonText,
+                           final String negativeButtonText, final Callback successCallback, Callback errorCallback) {
         Activity currentActivity = getCurrentActivity();
         if (currentActivity == null) {
-            // If getCurrentActivity is null, it could be because the app is backgrounded,
-            // so we show the dialog when the app resumes)
+
+            /* If getCurrentActivity is null, it could be because the app is backgrounded,
+            * so we show the dialog when the app resumes). */
             getReactApplicationContext().addLifecycleEventListener(new LifecycleEventListener() {
                 @Override
                 public void onHostResume() {
                     Activity currentActivity = getCurrentActivity();
                     if (currentActivity != null) {
                         getReactApplicationContext().removeLifecycleEventListener(this);
-                        showDialogInternal(title, message, button1Text, button2Text, successCallback, currentActivity);
+                        showDialogInternal(title, message, positiveButtonText, negativeButtonText, successCallback, currentActivity);
                     }
                 }
 
@@ -46,16 +69,24 @@ public class CodePushDialog extends ReactContextBaseJavaModule{
                 }
             });
         } else {
-            showDialogInternal(title, message, button1Text, button2Text, successCallback, currentActivity);
+            showDialogInternal(title, message, positiveButtonText, negativeButtonText, successCallback, currentActivity);
         }
     }
 
-    private void showDialogInternal(String title, String message, String button1Text,
-                                    String button2Text, final Callback successCallback, Activity currentActivity) {
+    /**
+     * Internal method for actually showing the dialog.
+     *
+     * @param title              title of the dialog.
+     * @param message            message to be displayed.
+     * @param positiveButtonText text on the positive action button.
+     * @param negativeButtonText test on the negative action button.
+     * @param successCallback    callback to handle "OK" events.
+     * @param currentActivity    application activity.
+     */
+    private void showDialogInternal(String title, String message, String positiveButtonText,
+                                    String negativeButtonText, final Callback successCallback, Activity currentActivity) {
         AlertDialog.Builder builder = new AlertDialog.Builder(currentActivity);
-
         builder.setCancelable(false);
-
         DialogInterface.OnClickListener clickListener = new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
@@ -68,37 +99,54 @@ public class CodePushDialog extends ReactContextBaseJavaModule{
                         successCallback.invoke(1);
                         break;
                     default:
-                        throw new CodePushUnknownException("Unknown button ID pressed.");
+                        CodePushLogUtils.trackException("Unknown button ID pressed.");
                 }
             }
         };
-
         if (title != null) {
             builder.setTitle(title);
         }
-
         if (message != null) {
             builder.setMessage(message);
         }
-
-        if (button1Text != null) {
-            builder.setPositiveButton(button1Text, clickListener);
+        if (positiveButtonText != null) {
+            builder.setPositiveButton(positiveButtonText, clickListener);
         }
-
-        if (button2Text != null) {
-            builder.setNegativeButton(button2Text, clickListener);
+        if (negativeButtonText != null) {
+            builder.setNegativeButton(negativeButtonText, clickListener);
         }
-
         try {
             AlertDialog dialog = builder.create();
             dialog.show();
         } catch (Exception e) {
-            CodePushRNUtils.log(e.toString());
+            CodePushLogUtils.trackException(e);
         }
     }
 
     @Override
     public String getName() {
         return "CodePushDialog";
+    }
+
+    @Override
+    public void shouldInstallUpdate(String title, String message, String acceptText, String declineText, final CodePushConfirmationCallback codePushConfirmationCallback) {
+        final Callback successCallback = new Callback() {
+            @Override
+            public void invoke(Object... args) {
+                String buttonNumber = args[0].toString();
+                if (buttonNumber.equals("0")) {
+                    codePushConfirmationCallback.onResult(true);
+                } else if (buttonNumber.equals("1")) {
+                    codePushConfirmationCallback.onResult(false);
+                }
+            }
+        };
+        final Callback errorCallback = new Callback() {
+            @Override
+            public void invoke(Object... args) {
+                codePushConfirmationCallback.throwError(new CodePushGeneralException("Exception occurred when showing a dialog. Args: " + Arrays.toString(args)));
+            }
+        };
+        showDialog(title, message, acceptText, declineText, successCallback, errorCallback);
     }
 }
