--- android/app/src/main/java/com/microsoft/codepush/react/datacontracts/CodePushUpdateResponse.java
@@ -1,8 +1,0 @@
-package com.microsoft.codepush.react.datacontracts;
-
-import com.google.gson.annotations.SerializedName;
-
-public class CodePushUpdateResponse {
-    @SerializedName("updateInfo")
-    public CodePushUpdateResponseUpdateInfo UpdateInfo;
-}
