--- android/app/src/main/java/com/microsoft/codepush/react/exceptions/CodePushInvalidPublicKeyException.java
@@ -1,12 +1,0 @@
-package com.microsoft.codepush.react.exceptions;
-
-public class CodePushInvalidPublicKeyException extends RuntimeException {
-
-    public CodePushInvalidPublicKeyException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public CodePushInvalidPublicKeyException(String message) {
-        super(message);
-    }
-}
