--- android/app/src/main/java/com/microsoft/codepush/react/exceptions/CodePushInvalidUpdateException.java
@@ -1,7 +1,0 @@
-package com.microsoft.codepush.react.exceptions;
-
-public class CodePushInvalidUpdateException extends RuntimeException {
-    public CodePushInvalidUpdateException(String message) {
-        super(message);
-    }
-}
