--- android/app/src/main/java/com/microsoft/codepush/react/exceptions/CodePushNotInitializedException.java
@@ -1,12 +1,0 @@
-package com.microsoft.codepush.react.exceptions;
-
-public final class CodePushNotInitializedException extends RuntimeException {
-
-    public CodePushNotInitializedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public CodePushNotInitializedException(String message) {
-        super(message);
-    }
-}
