--- android/app/src/main/java/com/microsoft/codepush/react/interfaces/CodePushBinaryVersionMismatchListener.java
@@ -1,7 +1,0 @@
-package com.microsoft.codepush.react.interfaces;
-
-import com.microsoft.codepush.react.datacontracts.CodePushRemotePackage;
-
-public interface CodePushBinaryVersionMismatchListener {
-    void binaryVersionMismatchChanged(CodePushRemotePackage update);
-}
