--- android/app/src/main/java/com/microsoft/codepush/react/interfaces/CodePushDownloadProgressListener.java
@@ -1,5 +1,0 @@
-package com.microsoft.codepush.react.interfaces;
-
-public interface CodePushDownloadProgressListener {
-    void downloadProgressChanged(long receivedBytes, long totalBytes);
-}
