--- android/app/src/main/java/com/microsoft/codepush/react/interfaces/CodePushSyncStatusListener.java
@@ -1,7 +1,0 @@
-package com.microsoft.codepush.react.interfaces;
-
-import com.microsoft.codepush.react.enums.CodePushSyncStatus;
-
-public interface CodePushSyncStatusListener {
-    void syncStatusChanged(CodePushSyncStatus syncStatus);
-}
