--- android/app/src/main/java/com/microsoft/codepush/react/interfaces/DownloadProgressCallback.java
@@ -1,7 +1,0 @@
-package com.microsoft.codepush.react.interfaces;
-
-import com.microsoft.codepush.react.DownloadProgress;
-
-public interface DownloadProgressCallback {
-    void call(DownloadProgress downloadProgress);
-}
