--- android/app/src/main/java/com/microsoft/codepush/reactv2/interfaces/ReactInstanceHolder.java
@@ -1,17 +1,0 @@
-package com.microsoft.codepush.reactv2.interfaces;
-
-import com.facebook.react.ReactInstanceManager;
-
-/**
- * Provides access to a {@link ReactInstanceManager}.
- * <p>
- * ReactNativeHost already implements this interface, if you make use of that react-native
- * component (just add `implements ReactInstanceHolder`).
- */
-public interface ReactInstanceHolder {
-
-    /**
-     * Get the current {@link ReactInstanceManager} instance. May return null.
-     */
-    ReactInstanceManager getReactInstanceManager();
-}
